+++ src/app/(main)/page.tsx
@@ -1,0 +1,18 @@
+'use client'
+import * as Button from "../components/Button/Button";
+import NavBar from "../components/NavBar/NavBar";
+import Counter from "../components/Counter/Counter";
+import Checkbox from "../components/CheckBox/Checkbox";
+import ProfilePhoto from "../components/ProfilePhoto/ProfilePhoto";
+import Information from "../(notUse)/information/page";
+import ModalContainer from "../components/Modals/ModalContainer";
+import TeacherMission from "../teacher-mission/page";
+import SidebarTeacher from "../components/SidebarTeacher/SidebarTeacher";
+import DifficultyButton from "../components/DifficultyButton/DifficultyButton";
+import ChoosenProfile from "../choosen-profile/page";
+
+export default function Home() {
+  return (
+    <></>
+  );
+}
